session4/pyproject.toml
[project]
name = "session3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "graphviz>=0.21",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "langsmith>=0.7.17",
    "python-dotenv>=1.2.2",
]
